Brandon G. Elzy - Refactored component
