+++ client/src/components/container/App/App.js
@@ -1,7 +1,8 @@
 import React, { Component } from "react";
 import { Sidebar } from "semantic-ui-react";
 import styled from "styled-components";
-import { CreateAlert, Topbar, VerticalSidebar } from '../../presentation/presentation.js';
+import { Topbar, VerticalSidebar } from '../../presentation/presentation.js';
+import CreateAlert from '../CreateAlert/CreateAlert.js';
 import "semantic-ui-css/semantic.min.css";
 
 export default class App extends Component {
